requirements.txt
pandas
google-genai
python-dotenv
streamlit